Probabilistic

ProbabilisticCircuits.Probabilistic.AbstractFlatMixture — Type

A probabilistic mixture model whose components are not themselves mixtures

ProbabilisticCircuits.Probabilistic.AbstractMetaMixture — Type

A probabilistic mixture model whose components are themselves mixtures

ProbabilisticCircuits.Probabilistic.AbstractMixture — Type

A probabilistic mixture model

ProbabilisticCircuits.Probabilistic.DisCache — Type

Cache pairwise / marginal distribution for all variables in one dataset

ProbabilisticCircuits.Probabilistic.FlatMixture — Type

A probabilistic mixture model of probabilistic circuits

ProbabilisticCircuits.Probabilistic.FlatMixtureWithFlow — Type

A mixture with cached flow circuits for each component (which are assumed to be ProbΔs)

ProbabilisticCircuits.Probabilistic.MetaMixture — Type

A probabilistic mixture model of mixture models

ProbabilisticCircuits.Probabilistic.aggregate_flows — Method

Compute and aggregate flows for mixture components

ProbabilisticCircuits.Probabilistic.blossom_bottom_up! — Method

Blossom bottom up method, vars are not used

ProbabilisticCircuits.Probabilistic.component_weights — Method

Get the component weights in this mixture

ProbabilisticCircuits.Probabilistic.components — Method

Get the components in this mixture

ProbabilisticCircuits.Probabilistic.conjoin_like — Method

Conjoin nodes in the same way as the example

ProbabilisticCircuits.Probabilistic.disjoin_like — Method

Disjoin nodes in the same way as the example

ProbabilisticCircuits.Probabilistic.init_mixture_with_flows — Method

Ensure we have a FlatMixtureWithFlow where the flow circuits have aggregate flow circuits as origin

ProbabilisticCircuits.Probabilistic.literal_like — Method

Construct a new literal node like the given node's type

ProbabilisticCircuits.Probabilistic.log_likelihood_per_instance — Method

Calculate log likelihood per instance for batches of samples.

ProbabilisticCircuits.Probabilistic.log_likelihood_per_instance — Method

Calculates log likelihood for a batch of fully observed samples. (Also retures the generated FlowΔ)

ProbabilisticCircuits.Probabilistic.log_likelihood_per_instance — Method

Calculate log likelihood for a batch of fully observed samples. (This is for when you already have a FlowΔ)

ProbabilisticCircuits.Probabilistic.log_likelihood_per_instance_component — Method

Log likelihood per instance and component. A vector of matrices per batch where the first dimension is instance, second is components.

ProbabilisticCircuits.Probabilistic.log_likelihood_per_instance_component — Method

Log likelihood per instance and component. First dimension is instance, second is components.

ProbabilisticCircuits.Probabilistic.log_likelihood_per_instance_component — Method

Log likelihood per instance and component. A vector of matrices per batch where the first dimension is instance, second is components.

ProbabilisticCircuits.Probabilistic.log_likelihood_per_instance_component — Method

Log likelihood per instance and component. First dimension is instance, second is components.

ProbabilisticCircuits.Probabilistic.marginal_log_likelihood_per_instance — Method

Calculate log likelihood for a batch of samples with partial evidence P(e). (Also returns the generated FlowΔ)

To indicate a variable is not observed, pass -1 for that variable.

ProbabilisticCircuits.Probabilistic.marginal_log_likelihood_per_instance — Method

Calculate log likelihood for a batch of samples with partial evidence P(e). (If you already have a FlowΔ)

To indicate a variable is not observed, pass -1 for that variable.

ProbabilisticCircuits.Probabilistic.metis_top_down — Method

Metis top down method

ProbabilisticCircuits.Probabilistic.mutual_information — Function

Calculate mutual information of given bit matrix bm, example weights w, and smoothing pseudocount α

ProbabilisticCircuits.Probabilistic.num_components — Method

Number of components in a mixture

ProbabilisticCircuits.Probabilistic.pr_constraint — Method

Calculate the probability of the logic formula given by sdd for the psdd

ProbabilisticCircuits.Probabilistic.prob_origin — Method

Return the first origin that is a probabilistic circuit

ProbabilisticCircuits.Probabilistic.prob_origin — Method

Return the first origin that is a probabilistic circuit node

ProbabilisticCircuits.Probabilistic.psdd_entropy — Method

Entropy of the distribution of the input psdd.

ProbabilisticCircuits.Probabilistic.psdd_kl_divergence — Method

KL divergence calculation for psdds that are not necessarily identical

ProbabilisticCircuits.Probabilistic.sample — Method

Sampling with Evidence from a psdd. Internally would call marginal pass up on a newly generated flow circuit.

ProbabilisticCircuits.Probabilistic.sample — Method

Sample from a PSDD without any evidence

ProbabilisticCircuits.Probabilistic.sample — Method

Sampling with Evidence from a psdd. Assuming already marginal pass up has been done on the flow circuit.

ProbabilisticCircuits.Probabilistic.test_bottom_up! — Method

Test bottom up method, split nodes by ascending order, balanced

ProbabilisticCircuits.Probabilistic.test_top_down — Method

Test top down method, split nodes by ascending order, balanced

ProbabilisticCircuits.Probabilistic.train_mixture — Method

Train a mixture model from data. Learning is initialized from the parameters stored in the given mixture. When a structure_learner is given, it will be called between EM steps to update circuit structures.

ProbabilisticCircuits.Probabilistic.train_mixture — Method

Train a mixture of probabilistic circuits from data, starting with random example weights.

ProbabilisticCircuits.Probabilistic.aggregate_flows_cached — Method

Aggregate already-computed flows for mixture components

ProbabilisticCircuits.Probabilistic.component_weights_per_example — Method

Compute the component weights for each example from likelihoods

ProbabilisticCircuits.Probabilistic.ensure_with_flows — Method

Convert a given flat mixture into one with cached flows

ProbabilisticCircuits.Probabilistic.log_likelihood_per_component_instance — Method

Log likelihood per component and instance. Outer vector is components, inner vector is instances

ProbabilisticCircuits.Probabilistic.log_likelihood_per_component_instance — Method

Log likelihood per component and instance. Outer vector is components, inner vector is instances

ProbabilisticCircuits.Probabilistic.mpe_simulate — Method

activesamples: bool vector indicating which samples are active for this node during mpe result: Matrix (numsamples, num_variables) indicating the final result of mpe

ProbabilisticCircuits.Probabilistic.random_example_weights — Method

Create random example weights that sum to one overall components

ProbabilisticCircuits.Probabilistic.set_mutual_information — Method

Calculate set mutual information

ProbabilisticCircuits.Probabilistic.weighted_batch_for_component — Method

Get a new weighted batch for this component