Reasoning

ProbabilisticCircuits.Reasoning.ExpFlowΔ — Method

Construct a upward expectation flow circuit from a given pair of PC and LC circuits Note that its assuming the two circuits share the same vtree

ProbabilisticCircuits.Reasoning.Expectation — Method

Missing values should be denoted by -1

ProbabilisticCircuits.Reasoning.DecoratorΔNodePair — Type

A expectation circuit node that has pair of origins of type PC and type LC

ProbabilisticCircuits.Reasoning.exp_f — Method

Has to be a Logistic⋁ with only one child, which is a leaf node

ProbabilisticCircuits.Reasoning.exp_fg — Method

Has to be a Logistic⋁ with only one child, which is a leaf node

ProbabilisticCircuits.Reasoning.moment_fg — Method

Calculating E[g^k * f]