Probabilistic Circuits

ProbabilisticCircuits.BitCircuitPair — Type

A bit circuit pair is a low-level representation of pairs of nodes traversed for a pair of probabilistic circuit logical circuit structure. For example, this is used for taking expectation of a Logistic/Regression circuit w.r.t. a probabilistic circuit.

The wiring of the circuit is captured by two matrices: nodes and elements.

Nodes are either leafs (input) or pair of Sum nodes

Elements are Pair of Product nodes

In addition, there is a vector of layers, where each layer is a list of node ids. Layer 1 is the leaf/input layer. Layer end is the circuit root.

And there is a vector of parents, pointing to element id parents of decision nodes.

Nodes are represented as a 6xN matrix where

nodes[1,:] is the first child pair id

nodes[2,:] is the last child pair id

nodes[3,:] is the first parent index

nodes[4,:] is the last parent index

nodes[5,:] is the id of corresponding node from the PC (first circuit)

nodes[6,:] is the id of corresponding node from the LC (first circuit)

Elements are represented by a 3xE matrix, where

elements[1,:] is the Product pair node id,

elements[2,:] is the (left,left) child node id

elements[3,:] is the (right right) child node id

Elements belonging to node pair i are elements[:, nodes[1,i]:nodes[2,i]]

Parents are represented as a one dimentional array Parents belonging to node pair i are parents[nodes[3,i]:nodes[4,i]]

ProbabilisticCircuits.CLT — Type

Chow-Liu Tree

ProbabilisticCircuits.ExpFlowCircuit — Method

Construct a upward expectation flow circuit from a given pair of PC and LC circuits Note that its assuming the two circuits share the same vtree

ProbabilisticCircuits.ExpFlowNode — Type

A expectation circuit node that has pair of origins of type PC and type LC

ProbabilisticCircuits.LogisticCircuit — Type

Root of the logistic circuit node hierarchy

ProbabilisticCircuits.LogisticInnerNode — Type

A logistic inner node

ProbabilisticCircuits.LogisticLeafNode — Type

A logistic leaf node

ProbabilisticCircuits.LogisticLiteralNode — Type

A logistic literal node

ProbabilisticCircuits.Logistic⋀Node — Type

A logistic conjunction node (And node)

ProbabilisticCircuits.Logistic⋁Node — Type

A logistic disjunction node (Or node)

ProbabilisticCircuits.NodePairId — Type

Integer identifier for a circuit pair node

ProbabilisticCircuits.NodePairIds — Type

The BitCircuitPair ids associated with a node

ProbabilisticCircuits.ParamBitCircuit — Type

A BitCircuit with parameters attached to the elements

ProbabilisticCircuits.PlainMulNode — Type

A probabilistic conjunction node (multiplication node)

ProbabilisticCircuits.PlainProbCircuit — Type

Root of the plain probabilistic circuit node hierarchy

ProbabilisticCircuits.PlainProbInnerNode — Type

A probabilistic inner node

ProbabilisticCircuits.PlainProbLeafNode — Type

A probabilistic leaf node

ProbabilisticCircuits.PlainProbLiteralNode — Type

A probabilistic literal node

ProbabilisticCircuits.PlainSumNode — Type

A probabilistic disjunction node (summation node)

ProbabilisticCircuits.ProbCircuit — Type

Root of the probabilistic circuit node hierarchy

ProbabilisticCircuits.ProbCircuit — Method

Evaluate a probabilistic circuit as a function

ProbabilisticCircuits.SharedMulNode — Type

A shared probabilistic multiplcation node

ProbabilisticCircuits.SharedProbCircuit — Type

Root of the shared probabilistic circuit node hierarchy

ProbabilisticCircuits.SharedProbInnerNode — Type

A shared probabilistic inner node

ProbabilisticCircuits.SharedProbLeafNode — Type

A shared probabilistic leaf node

ProbabilisticCircuits.SharedProbLiteralNode — Type

A shared probabilistic literal node

ProbabilisticCircuits.SharedSumNode — Type

A shared probabilistic summation node

ProbabilisticCircuits.StructMulNode — Type

A plain structured probabilistic conjunction node

ProbabilisticCircuits.StructProbCircuit — Type

Root of the plain structure probabilistic circuit node hierarchy

ProbabilisticCircuits.StructProbInnerNode — Type

A plain structured probabilistic inner node

ProbabilisticCircuits.StructProbLeafNode — Type

A plain structured probabilistic leaf node

ProbabilisticCircuits.StructProbLiteralNode — Type

A plain structured probabilistic literal leaf node, representing the positive or negative literal of its variable

ProbabilisticCircuits.StructSumNode — Type

A plain structured probabilistic disjunction node

LogicCircuits.Utils.print_tree — Method

Print edges and vertices of a ChowLiu tree

LogicCircuits.vtree — Method

Get the vtree corresponding to the argument, or nothing if the node has no vtree

ProbabilisticCircuits.EVI — Function

EVI(pc, data)

Computes the log likelihood data given full evidence. Outputs $\log{p(x)}$ for each datapoint.

ProbabilisticCircuits.Expectation — Method

Expectation(pc::ProbCircuit, lc::LogisticCircuit, data)

Compute Expected Prediction of a Logistic/Regression Circuit w.r.t to a ProbabilistcCircuit

Missing values should be denoted by missing See: On Tractable Computation of Expected Predictions arxiv.org/abs/1910.02182

ProbabilisticCircuits.ExpectationBit — Method

ExpectationBit(pc::ProbCircuit, lc::LogisticCircuit, data)
ExpectationBit(pbc::ParamBitCircuitPair, pc::ProbCircuit, lc::LogisticCircuit, data)

Compute Expected Prediction of a Logistic/Regression Circuit w.r.t to a ProbabilistcCircuit. Uses the BitCircuitPair implementation, which enables parrallism on both CPU/GPU. If the data is on gpu the gpu version will get called (repectively for cpu.)

Missing values should be denoted by missing See: On Tractable Computation of Expected Predictions arxiv.org/abs/1910.02182

ProbabilisticCircuits.ExpectationUpward — Method

Expected Prediction of LC w.r.t PC. This implementation uses the computation graph approach.

ProbabilisticCircuits.MAP — Function

Maximum a-posteriori queries

ProbabilisticCircuits.MAR — Function

MAR(pc, data)

Computes Marginal log likelhood of data. See docs for marginal.

ProbabilisticCircuits.Moment — Method

Moment(pc::ProbCircuit, lc::LogisticCircuit, data, moment::Int)

Compute higher moments of Expected Prediction for the pair of Logistic/Regression Circuit, ProbabilistcCircuit

ProbabilisticCircuits.apply_entropy_reg_cpu — Method

Add entropy regularization to a deterministic (see LogicCircuits.isdeterministic) probabilistic circuit. alpha is a hyperparameter that balances the weights between the likelihood and the entropy: rgmax{ heta} L( heta) = llmean( heta) + alpha * entropy( heta).

ProbabilisticCircuits.compile_canonical_literals — Method

Construct literal nodes given variable var

ProbabilisticCircuits.compile_decision_node — Method

Construct decision nodes given primes and subs

ProbabilisticCircuits.compile_decision_nodes — Method

Construct decision nodes conditiond on different distribution

ProbabilisticCircuits.compile_sdd_from_clt — Method

Compile a psdd circuit from clt and vtree

ProbabilisticCircuits.compile_true_nodes — Method

Construct true nodes given variable var

ProbabilisticCircuits.eFlow — Method

Pick the edge with maximum flow

ProbabilisticCircuits.eRand — Method

Pick the edge randomly

ProbabilisticCircuits.estimate_parameters — Method

Maximum likilihood estimation of parameters given data

use_gpu: If set to true, use gpu learning no matter which device data is in.

bagging support: If pc is a SharedProbCircuit and data is an array of DataFrames with the same number of "components", learn each circuit with its corresponding dataset.

ProbabilisticCircuits.estimate_parameters_em — Method

Expectation maximization parameter learning given missing data

ProbabilisticCircuits.exp_f — Method

Has to be a Logistic⋁Node with only one child, which is a leaf node

ProbabilisticCircuits.exp_fg — Method

Has to be a Logistic⋁Node with only one child, which is a leaf node

ProbabilisticCircuits.get_cpt — Method

Calculate CPT of child conditioned on parent from dis_cache

ProbabilisticCircuits.init_expectations — Method

init_expectations(circuit::ParamBitCircuitPair, data, reuse_f, reuse_g, nodes_num, classes_num; Float = Float32)

Initialized the input layer for `fvalues` and `gvalues`. 
To reduce memory allocation can pass `reuse_f` and `reuse_g` to get reused (very useful during batching).

ProbabilisticCircuits.init_marginal — Method

Initialize values from the data (data frames)

ProbabilisticCircuits.ismul — Method

Is the node a multiplication?

ProbabilisticCircuits.issum — Method

Is the node a summation?

ProbabilisticCircuits.kl_divergence — Function

Calculate KL divergence calculation for pcs that are not necessarily identical

ProbabilisticCircuits.learn_chow_liu_tree — Method

learn a Chow-Liu tree from training set train_x, with Laplace smoothing factor α, specifying the tree root by clt_root return a CLT

ProbabilisticCircuits.learn_chow_liu_tree_circuit — Method

Learning from data a structured-decomposable circuit with several structure learning algorithms

ProbabilisticCircuits.learn_circuit — Method

Learn structure of a single structured decomposable circuit

ProbabilisticCircuits.learn_circuit_miss — Method

Learn structure of a single structured decomposable circuit from missing data.

Missing feature are denoted by missing. Median Imputation is used by default for initial structure (set impute_method for other options).

ProbabilisticCircuits.learn_circuit_mixture — Method

Given a circuit, learns a mixture of structure decomposable circuits based on that circuit

learn_circuit_mixture(pc, data; num_mix=5, pseudocount=1.0, maxiter=20)

ProbabilisticCircuits.learn_strudel — Method

learn_strudel(train_x; num_mix = 5, init_maxiter = 10, em_maxiter=20)

Learn a mixture of circuits See "Strudel: Learning Structured-Decomposable Probabilistic Circuits. arxiv.org/abs/2007.09331

ProbabilisticCircuits.learn_vtree_from_clt — Method

Learn a vtree from clt, with strategy (close to) linear or balanced

ProbabilisticCircuits.log_likelihood — Method

log_likelihood(pc, data)

Compute the likelihood of the PC given the data

ProbabilisticCircuits.log_likelihood_avg — Method

log_likelihood_avg(pc, data)

Compute the likelihood of the PC given the data, averaged over all instances in the data

ProbabilisticCircuits.log_likelihood_per_instance — Function

Compute the likelihood of the PC given each individual instance in the data

ProbabilisticCircuits.map_child — Method

Find the MAP child value and node id of a given decision node

ProbabilisticCircuits.marginal — Method

marginal(root::ProbCircuit, data::Union{Real,Missing}...)
marginal(root::ProbCircuit, data::Union{Vector{Union{Bool,Missing}},CuVector{UInt8}})
marginal(circuit::ProbCircuit, data::DataFrame)
marginal(circuit::ParamBitCircuit, data::DataFrame)::AbstractVector
marginal(circuit::SharedProbCircuit, data::DataFrame, weights::Union{AbstractArray, Nothing}; component_idx)

Evaluate marginals of the circuit bottom-up for given input(s).

Missing values should be denoted by missing in the data.

Outputs $\log{p(x^o)}$ for each data point.

ProbabilisticCircuits.marginal_all — Method

Evaluate the probabilistic circuit bottom-up for a given input and return the marginal probability value of all nodes

ProbabilisticCircuits.marginal_flows — Function

Compute the marginal and flow of each node

ProbabilisticCircuits.marginal_flows_down — Function

When marginals of nodes have already been computed, do a downward pass computing the marginal flows at each node

ProbabilisticCircuits.marginal_flows_down_layers — Method

Evaluate marginals of the layers of a parameter bit circuit on the CPU (SIMD & multi-threaded)

ProbabilisticCircuits.marginal_flows_down_layers — Method

Pass marginal flows down the layers of a bit circuit on the GPU

ProbabilisticCircuits.marginal_flows_down_layers_cuda — Method

CUDA kernel for passing marginal flows down circuit

ProbabilisticCircuits.marginal_layers — Method

Compute marginals on the CPU (SIMD & multi-threaded)

ProbabilisticCircuits.marginal_layers — Method

Compute marginals on the GPU

ProbabilisticCircuits.marginal_layers_cuda — Method

CUDA kernel for circuit evaluation

ProbabilisticCircuits.marginal_log_likelihood — Method

marginal_log_likelihood(pc, data)

Compute the marginal likelihood of the PC given the data

ProbabilisticCircuits.marginal_log_likelihood_avg — Method

Compute the marginal likelihood of the PC given the data, averaged over all instances in the data

ProbabilisticCircuits.max_a_posteriori — Method

max_a_posteriori(root::ProbCircuit, data::Union{Bool,Missing}...)
max_a_posteriori(root::ProbCircuit, data::Union{Vector{<:Union{Bool,Missing}},CuVector{UInt8}})
max_a_posteriori(circuit::ProbCircuit, data::DataFrame)
max_a_posteriori(pbc::ParamBitCircuit, data; Float=Float32)

Evaluate maximum a-posteriori state of the circuit for given input(s).

Outputs the states, and the corresponding probabilities (in log domain).

ProbabilisticCircuits.metis_top_down — Method

Metis top down method

ProbabilisticCircuits.mode — Function

Mode of the distribution

ProbabilisticCircuits.moment_fg — Method

Calculating E[g^k * f]

ProbabilisticCircuits.mul_nodes — Method

Get the list of multiplication nodes in a given circuit

ProbabilisticCircuits.multiply — Function

Multiply nodes into a single circuit

ProbabilisticCircuits.num_components — Method

How many components are mixed together in this shared circuit?

ProbabilisticCircuits.num_parameters — Method

Count the number of parameters in the circuit

ProbabilisticCircuits.num_parameters_node — Method

Count the number of parameters in the node

ProbabilisticCircuits.params — Method

Get the parameters associated with a or node

ProbabilisticCircuits.params — Method

Get the parameters associated with a sum node

ProbabilisticCircuits.parent_vector — Method

Get parent vector of a tree

ProbabilisticCircuits.pr_constraint — Function

Calculate the probability of the logic formula given by LC for the PC

ProbabilisticCircuits.sample — Method

sample(pc::ProbCircuit, num_samples)
sample(pc::ProbCircuit, num_samples, evidences)

Sample states from the probabilistic circuit distribution. Also can do conditional sampling if evidence is given (any subset of features).

ProbabilisticCircuits.sum_nodes — Method

Get the list of summation nodes in a given circuit

ProbabilisticCircuits.summate — Function

Sum nodes into a single circuit

ProbabilisticCircuits.to_sampled_dataframes — Method

Convert an array of samples into a vector of dataframes

ProbabilisticCircuits.uniform_parameters — Method

Uniform distribution

ProbabilisticCircuits.vMI — Method

Pick the variable with maximum sum of mutual information

ProbabilisticCircuits.vRand — Method

Pick the variable randomly

ProbabilisticCircuits.wmc_chavira — Function

wmc_chavira(root::LogicCircuit; varprob::Function)::Float

Compute weighted model count in the context of a chavira encoding (default negative literals to 1) Probability of each variable is given by varprob Function which defauls to 1/2 for every variable.