Probabilistic Circuits

ProbabilisticCircuits.Categorical — Type

A N-value categorical input distribution ranging over integers [0...N-1]

ProbabilisticCircuits.FlatVectors — Type

An isbits representation of a AbstractVector{<:AbstractVector}

ProbabilisticCircuits.Indicator — Type

A input distribution node that places all probability on a single value

ProbabilisticCircuits.Literal — Type

A logical literal input distribution node

ProbabilisticCircuits.NodeType — Type

Probabilistic circuit node types

ProbabilisticCircuits.NodeType — Method

Get the probabilistic circuit node type

ProbabilisticCircuits.PlainInnerNode — Type

A probabilistic inner node

ProbabilisticCircuits.PlainInputNode — Type

A probabilistic input node

ProbabilisticCircuits.PlainMulNode — Type

A probabilistic multiplication node

ProbabilisticCircuits.PlainProbCircuit — Type

Root of the plain probabilistic circuit node hierarchy

ProbabilisticCircuits.PlainSumNode — Type

A probabilistic summation node

ProbabilisticCircuits.ProbCircuit — Type

Root of the probabilistic circuit node hierarchy

ProbabilisticCircuits.RegionGraph — Type

Root of region graph node hierarchy

Base.read — Method

Base.read(file::AbstractString, ::Type{C}) where C <: ProbCircuit

Reads circuit from file; uses extension to detect format type, for example ".psdd" for PSDDs.

Base.write — Method

Base.write(file::AbstractString, circuit::ProbCircuit)

Writes circuit to file; uses file name extention to detect file format.

DirectedAcyclicGraphs.foldup — Method

foldup(node::ProbCircuit, 
    f_i::Function, 
    f_m::Function, 
    f_s::Function)::T where {T}

Compute a function bottom-up on the circuit. f_in is called on input nodes, f_m is called on product nodes, and f_s is called on sum nodes. Values of type T are passed up the circuit and given to f_m and f_s through a callback from the children.

DirectedAcyclicGraphs.foldup_aggregate — Method

foldup_aggregate(node::ProbCircuit, 
    f_i::Function, 
    f_m::Function, 
    f_s::Function, 
    ::Type{T})::T where T

Compute a function bottom-up on the circuit. f_in is called on input nodes, f_m is called on product nodes, and f_s is called on sum nodes. Values of type T are passed up the circuit and given to f_m and f_s in an aggregate vector from the children.

ProbabilisticCircuits.MAP — Method

MAP(pc::ProbCircuit, data::Matrix; batch_size, Float=Float32)

Evaluate max a posteriori (MAP) state of the circuit for given input(s) on cpu.

Note: This algorithm is only exact if the circuit is both decomposable and determinisitic. If the circuit is only decomposable and not deterministic, this will give inexact results without guarantees.

ProbabilisticCircuits.balance_sum — Method

Makes sure the sum nodes does not have too many children. Makes balanced sums of sums to reduce children count.

ProbabilisticCircuits.balanced_fully_factorized_leaves — Method

Makes sure input nodes don't have too many parents. Makes a dummy sum node for each input per partition. Then nodes corresponding to the partition use the dummy node as their children instead of the input node. This way instead of numnodesroot * numnodesleaf, we would have numnodesroot parents nodes.

ProbabilisticCircuits.cleanup_memory — Method

Cleansup allocated memory. Used internally.

ProbabilisticCircuits.dist — Function

Get the distribution of a PC input node

ProbabilisticCircuits.eval_circuit! — Method

eval_circuit!(mars, linPC::AbstractVector{<:ProbCircuit}, data::Matrix, example_ids;  node2idx::Dict{ProbCircuit, UInt32}, Float=Float32)

Used internally. Evaluates the marginals of the circuit on cpu. Stores the values in mars.

mars: (batch_size, nodes)

linPC: linearized PC. (i.e. linearize(pc))

data: data Matrix (num_examples, features)

example_ids: Array or collection of ids for current batch

node2idx: Index of each ProbCircuit node in the linearized circuit

ProbabilisticCircuits.eval_circuit_max! — Method

eval_circuit_max!(mars, linPC::AbstractVector{<:ProbCircuit}, data::Matrix, example_ids;  node2idx::Dict{ProbCircuit, UInt32}, Float=Float32)

Used internally. Evaluates the MAP upward pass of the circuit on cpu. Stores the values in mars.

mars: (batch_size, nodes)

linPC: linearized PC. (i.e. linearize(pc))

data: data Matrix (num_examples, features)

example_ids: Array or collection of ids for current batch

node2idx: Index of each ProbCircuit node in the linearized circuit

ProbabilisticCircuits.full_batch_em — Method

full_batch_em(bpc::CuBitsProbCircuit, raw_data::CuArray, num_epochs; batch_size, pseudocount)

Update the paramters of the CuBitsProbCircuit by doing EM on the full batch (i.e. update paramters at the end of each epoch).

ProbabilisticCircuits.hclt — Method

hclt(data, num_hidden_cats; num_cats = nothing, input_type = LiteralDist)

Learns HiddenChowLiuTree (hclt) circuit structure from data.

data: Matrix or CuMatrix

num_hidden_cats: Number of categories in hidden variables

input_type: Distribution type for the inputs

num_cats: Number of categories (in case of categorical inputs). Automatically deduced if not given explicilty.

ProbabilisticCircuits.init_parameters — Method

init_parameters(pc::ProbCircuit; perturbation = 0.0)

Initialize parameters of ProbCircuit.

ProbabilisticCircuits.inputnodes — Method

Get all input nodes in a given circuit

ProbabilisticCircuits.inputs — Method

Get the inputs of a PC node

ProbabilisticCircuits.isinput — Method

Is the node an input node?

ProbabilisticCircuits.ismul — Method

Is the node a multiplication?

ProbabilisticCircuits.isonlysubedge — Method

whether this sub edge is the only outgoing edge from sub

ProbabilisticCircuits.ispartial — Method

whether this series of edges is partial or complete

ProbabilisticCircuits.issum — Method

Is the node a summation?

ProbabilisticCircuits.loglikelihood — Method

loglikelihood(root::ProbCircuit, data::Matrix, example_id; Float=Float32)

Computes marginal loglikelihood recursively on cpu for a single instance data[example_id, :].

Note: Quite slow, only use for demonstration/educational purposes.

ProbabilisticCircuits.loglikelihood — Method

loglikelihood(bpc::CuBitsProbCircuit, data::CuArray; batch_size, mars_mem = nothing)

Computes Average loglikelihood of circuit given the data using gpu. See loglikelihoods for more details.

ProbabilisticCircuits.loglikelihoods — Method

loglikelihoods(pc::ProbCircuit, data::Matrix)

Computes loglikelihoods of the circuit over the data on cpu. Linearizes the circuit and computes the marginals in batches.

ProbabilisticCircuits.loglikelihoods — Method

loglikelihoods(bpc::CuBitsProbCircuit, data::CuArray; batch_size, mars_mem = nothing)

Returns loglikelihoods for each datapoint on gpu. Missing values should be denoted by missing.

bpc: BitCircuit on gpu

data: CuArray{Union{Missing, data_types...}}

batch_size

mars_mem: Not required, advanced usage. CuMatrix to reuse memory and reduce allocations. See prep_memory and cleanup_memory.

ProbabilisticCircuits.loglikelihoods_vectorized — Method

Note: Experimental**; will be removed or renamed later

ProbabilisticCircuits.map_down_rec! — Method

map_down_rec!(mars, node::ProbCircuit, data, states::Matrix, batch_idx, example_idx; node2idx::Dict{ProbCircuit, UInt32}, Float=Float32)

Downward pass on cpu for MAP. Recursively chooses the best (max) sum node children according to the "MAP upward pass" values. Updates the missing values with map_state of that input node.

ProbabilisticCircuits.mini_batch_em — Method

mini_batch_em(bpc::CuBitsProbCircuit, raw_data::CuArray, num_epochs; batch_size, pseudocount, 
    param_inertia, param_inertia_end = param_inertia, shuffle=:each_epoch)

Update the parameters of the CuBitsProbCircuit by doing EM, update the parameters after each batch.

ProbabilisticCircuits.mulnodes — Method

Get all multiplication nodes in a given circuit

ProbabilisticCircuits.multiply — Function

Multiply nodes into a single circuit

ProbabilisticCircuits.num_inputs — Method

Number of inputs of a PC node

ProbabilisticCircuits.num_parameters — Function

Count the number of parameters in the circuit

ProbabilisticCircuits.num_parameters_node — Method

Count the number of parameters in the node

ProbabilisticCircuits.num_randvars — Method

Number of variables in the data structure

ProbabilisticCircuits.params — Method

Get the parameters associated with a node

ProbabilisticCircuits.prep_memory — Function

prep_memory(reuse, sizes, exact = map(x -> true, sizes))

Mostly used internally. Prepares memory for the specifed size, reuses reuse if possible to avoid memory allocation/deallocation.

ProbabilisticCircuits.psdd_num_nodes_leafs — Method

Count the number of decision and leaf nodes in the PSDD

ProbabilisticCircuits.random_region_graph — Method

random_region_graph(X::AbstractVector, depth::Int = 5, replicas::Int = 2, num_splits::Int = 2)

X: Vector of all variables to include; for the root region

depth: how many layers to do splits

replicas: number of replicas or paritions (replicas only used for the root region; for other regions only 1 parition (inner nodes), or 0 parition for leaves)

num_splits: number of splits for each parition; split variables into random equaly sized regions

ProbabilisticCircuits.randvars — Function

variables(pc::ProbCircuit)::BitSet

Get a bitset of variables mentioned in the circuit.

ProbabilisticCircuits.region_graph_2_pc — Method

region_graph_2_pc(node::RegionGraph; num_nodes_root, num_nodes_region, num_nodes_leaf, balance_childs_parents)

num_nodes_root: number of sum nodes in the root region

num_nodes_leaf: number of sum nodes per leaf region

num_nodes_region: number of in each region except root and leaves

ProbabilisticCircuits.sample — Method

sample(pc::ProbCircuit, num_samples; rng = default_rng())

Generate num_samples from the joint distribution of the circuit conditined on the data.

ProbabilisticCircuits.sample — Method

sample(pc::ProbCircuit, num_samples; rng = default_rng())

Generate num_samples from the joint distribution of the circuit without any conditions.

ProbabilisticCircuits.soften_data — Method

Turn binary data into floating point data close to 0 and 1.

ProbabilisticCircuits.summate — Function

Sum nodes into a single circuit

ProbabilisticCircuits.sumnodes — Method

Get all summation nodes in a given circuit

ProbabilisticCircuits.update_parameters — Method

map parameters from BitsPC back to the ProbCircuit it was created from