Utils

ProbabilisticCircuits.Utils — Module

Module with general utilities and missing standard library features that could be useful in any Julia project

ProbabilisticCircuits.Utils.DisCache — Type

Cache pairwise / marginal distribution for all variables in one dataset

ProbabilisticCircuits.Utils.bernoulli_partition — Method

Split X into two partitions A and B, where A is a Bernoulli sample of each element in X with probability p and B=X∖A. Guarantees at least one element in A and B.

ProbabilisticCircuits.Utils.generate_all — Method

Given some missing values generates all possible fillings

ProbabilisticCircuits.Utils.generate_data_all — Method

Generates all possible binary configurations of size N

ProbabilisticCircuits.Utils.kfold — Method

Returns a(n index) partitioning a la k-fold.

ProbabilisticCircuits.Utils.mutual_information — Function

Calculate mutual information of given bit matrix bm, example weights w, and smoothing pseudocount α

ProbabilisticCircuits.Utils.one_hot_encode — Method

One-hot encode data (2-D Array) based on categories (1-D Array) Each row of the return value is a concatenation of one-hot encoding of elements of the same row in data Assumption: both input arrays have elements of same type

ProbabilisticCircuits.Utils.pairwise_marginals — Function

Compute an array giving all pairwise marginals estimated on empirical (weighted) data

ProbabilisticCircuits.Utils.sample_vtree — Method

Samples a Vtree with a right bias of p. If p<0, then uniformly sample vtrees.

ProbabilisticCircuits.Utils.set_mutual_information — Method

Calculate set mutual information