Public APIs

This page lists documentation for the most commonly used public APIs of ProbabilisticCircuits.jl. Visit the internals section for a auto generated documentation for more public API and internal APIs.

Public APIs

Loading Circuits

Saving Interface

Learning Circuits

Circuit Queries

Compilation

Loading Circuits

Missing docstring.

Missing docstring for load_prob_circuit. Check Documenter's build log for details.

Missing docstring.

Missing docstring for load_struct_prob_circuit. Check Documenter's build log for details.

Missing docstring.

Missing docstring for load_logistic_circuit. Check Documenter's build log for details.

Saving Interface

Missing docstring.

Missing docstring for save_circuit. Check Documenter's build log for details.

Missing docstring.

Missing docstring for save_as_psdd. Check Documenter's build log for details.

Missing docstring.

Missing docstring for save_as_logistic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for save_as_dot. Check Documenter's build log for details.

Learning Circuits

ProbabilisticCircuits.learn_circuit — Function

Learn structure of a single structured decomposable circuit

ProbabilisticCircuits.learn_strudel — Function

learn_strudel(train_x; num_mix = 5, init_maxiter = 10, em_maxiter=20)

Learn a mixture of circuits See "Strudel: Learning Structured-Decomposable Probabilistic Circuits. arxiv.org/abs/2007.09331

ProbabilisticCircuits.estimate_parameters — Function

Maximum likilihood estimation of parameters given data

use_gpu: If set to true, use gpu learning no matter which device data is in.

bagging support: If pc is a SharedProbCircuit and data is an array of DataFrames with the same number of "components", learn each circuit with its corresponding dataset.

Missing docstring.

Missing docstring for learn_parameters. Check Documenter's build log for details.

ProbabilisticCircuits.learn_chow_liu_tree_circuit — Function

Learning from data a structured-decomposable circuit with several structure learning algorithms

Circuit Queries

ProbabilisticCircuits.marginal — Function

marginal(root::ProbCircuit, data::Union{Real,Missing}...)
marginal(root::ProbCircuit, data::Union{Vector{Union{Bool,Missing}},CuVector{UInt8}})
marginal(circuit::ProbCircuit, data::DataFrame)
marginal(circuit::ParamBitCircuit, data::DataFrame)::AbstractVector
marginal(circuit::SharedProbCircuit, data::DataFrame, weights::Union{AbstractArray, Nothing}; component_idx)

Evaluate marginals of the circuit bottom-up for given input(s).

Missing values should be denoted by missing in the data.

Outputs $\log{p(x^o)}$ for each data point.

ProbabilisticCircuits.max_a_posteriori — Function

max_a_posteriori(root::ProbCircuit, data::Union{Bool,Missing}...)
max_a_posteriori(root::ProbCircuit, data::Union{Vector{<:Union{Bool,Missing}},CuVector{UInt8}})
max_a_posteriori(circuit::ProbCircuit, data::DataFrame)
max_a_posteriori(pbc::ParamBitCircuit, data; Float=Float32)

Evaluate maximum a-posteriori state of the circuit for given input(s).

Outputs the states, and the corresponding probabilities (in log domain).

ProbabilisticCircuits.sample — Function

sample(pc::ProbCircuit, num_samples)
sample(pc::ProbCircuit, num_samples, evidences)

Sample states from the probabilistic circuit distribution. Also can do conditional sampling if evidence is given (any subset of features).

Missing docstring.

Missing docstring for Expectation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Moment. Check Documenter's build log for details.

Compilation

ProbabilisticCircuits.compile_sdd_from_clt — Function

Compile a psdd circuit from clt and vtree