Public APIs

This page lists documentation for the most commonly used public APIs of ProbabilisticCircuits.jl. Visit the internals section for a auto generated documentation for more public API and internal APIs.

Public APIs

Loading Circuits

Saving Interface

Learning Circuits

Circuit Queries

Compilation

Loading Circuits

ProbabilisticCircuits.LoadSave.load_prob_circuit — Function

Load a probabilistic circuit from file. Support circuit file formats: * ".psdd" for PSDD files

ProbabilisticCircuits.LoadSave.load_struct_prob_circuit — Function

Load a structured probabilistic circuit from file. Support circuit file formats: * ".psdd" for PSDD files Supported vtree file formats: * ".vtree" for Vtree files

ProbabilisticCircuits.LoadSave.load_logistic_circuit — Function

Load a logistic circuit from file. Support circuit file formats: * ".circuit" for logistic files Supported vtree file formats: * ".vtree" for Vtree files

Saving Interface

LogicCircuits.LoadSave.save_circuit — Function

Save a circuit to file

ProbabilisticCircuits.LoadSave.save_as_psdd — Function

Save a circuit to PSDD file format

ProbabilisticCircuits.LoadSave.save_as_logistic — Function

Save a logistic circuit to file

LogicCircuits.LoadSave.save_as_dot — Function

Save prob circuits to .dot file

Learning Circuits

ProbabilisticCircuits.learn_circuit — Function

Learn structure of a single structured decomposable circuit

ProbabilisticCircuits.learn_strudel — Function

learn_strudel(train_x; num_mix = 5, init_maxiter = 10, em_maxiter=20)

Learn a mixture of circuits See "Strudel: Learning Structured-Decomposable Probabilistic Circuits. arxiv.org/abs/2007.09331

ProbabilisticCircuits.estimate_parameters — Function

Maximum likilihood estimation of parameters given data

use_gpu: If set to true, use gpu learning no matter which device data is in.

bagging support: If pc is a SharedProbCircuit and data is an array of DataFrames with the same number of "components", learn each circuit with its corresponding dataset.

ProbabilisticCircuits.Logistic.learn_parameters — Function

LogisticCircuit Parameter learning through gradient descents Note: data need to be DataFrame and Labels need to be in one-hot form.

ProbabilisticCircuits.learn_chow_liu_tree_circuit — Function

Learning from data a structured-decomposable circuit with several structure learning algorithms

Circuit Queries

ProbabilisticCircuits.marginal — Function

marginal(root::ProbCircuit, data::Union{Real,Missing}...)
marginal(root::ProbCircuit, data::Union{Vector{Union{Bool,Missing}},CuVector{UInt8}})
marginal(circuit::ProbCircuit, data::DataFrame)
marginal(circuit::ParamBitCircuit, data::DataFrame)::AbstractVector
marginal(circuit::SharedProbCircuit, data::DataFrame, weights::Union{AbstractArray, Nothing}; component_idx)

Evaluate marginals of the circuit bottom-up for given input(s).

Missing values should be denoted by missing in the data.

Outputs $\log{p(x^o)}$ for each data point.

ProbabilisticCircuits.max_a_posteriori — Function

max_a_posteriori(root::ProbCircuit, data::Union{Bool,Missing}...)
max_a_posteriori(root::ProbCircuit, data::Union{Vector{<:Union{Bool,Missing}},CuVector{UInt8}})
max_a_posteriori(circuit::ProbCircuit, data::DataFrame)
max_a_posteriori(pbc::ParamBitCircuit, data; Float=Float32)

Evaluate maximum a-posteriori state of the circuit for given input(s).

Outputs the states, and the corresponding probabilities (in log domain).

ProbabilisticCircuits.sample — Function

sample(pc::ProbCircuit, num_samples)
sample(pc::ProbCircuit, num_samples, evidences)

Sample states from the probabilistic circuit distribution. Also can do conditional sampling if evidence is given (any subset of features).

ProbabilisticCircuits.Expectation — Function

Expectation(pc::ProbCircuit, lc::LogisticCircuit, data)

Compute Expected Prediction of a Logistic/Regression Circuit w.r.t to a ProbabilistcCircuit

Missing values should be denoted by missing See: On Tractable Computation of Expected Predictions arxiv.org/abs/1910.02182

ProbabilisticCircuits.Moment — Function

Moment(pc::ProbCircuit, lc::LogisticCircuit, data, moment::Int)

Compute higher moments of Expected Prediction for the pair of Logistic/Regression Circuit, ProbabilistcCircuit

Compilation

ProbabilisticCircuits.compile_sdd_from_clt — Function

Compile a psdd circuit from clt and vtree